+++ sdk/formrecognizer/Azure.AI.FormRecognizer/src/CopyModelOperation.cs
@@ -41,6 +41,8 @@ namespace Azure.AI.FormRecognizer.Training
         /// <summary>An ID representing the operation that can be used along with <see cref="_modelId"/> to poll for the status of the long-running operation.</summary>
         private readonly string _resultId;
 
+        private RequestFailedException _requestFailedException;
+
         /// <inheritdoc/>
         public override string Id { get; }
 
@@ -51,7 +53,7 @@ namespace Azure.AI.FormRecognizer.Training
             {
                 if (HasCompleted && !HasValue)
 #pragma warning disable CA1065 // Do not raise exceptions in unexpected locations
-                    throw new InvalidOperationException("Copy model operation failed");
+                    throw _requestFailedException;
 #pragma warning restore CA1065 // Do not raise exceptions in unexpected locations
                 else
                     return OperationHelpers.GetValue(ref _value);
@@ -164,14 +166,17 @@ namespace Azure.AI.FormRecognizer.Training
 
                 if (update.Value.Status == OperationStatus.Succeeded)
                 {
-                    _hasCompleted = true;
+                    // we need to first assign a vaue and then mark the operation as completed to avoid race conditions
                     _value = ConvertValue(update.Value, _targetModelId, CustomFormModelStatus.Ready);
+                    _hasCompleted = true;
                 }
                 else if (update.Value.Status == OperationStatus.Failed)
                 {
+                    _requestFailedException = await ClientCommon
+                        .CreateExceptionForFailedOperationAsync(async, _diagnostics, _response, update.Value.CopyResult.Errors)
+                        .ConfigureAwait(false);
                     _hasCompleted = true;
-
-                    throw await CreateExceptionForFailedOperationAsync(async, update.Value.CopyResult.Errors).ConfigureAwait(false);
+                    throw _requestFailedException;
                 }
             }
 
@@ -185,37 +190,6 @@ namespace Azure.AI.FormRecognizer.Training
                 result.CreatedDateTime,
                 result.LastUpdatedDateTime,
                 status);
-        }
-
-        private async ValueTask<RequestFailedException> CreateExceptionForFailedOperationAsync(bool async, IReadOnlyList<FormRecognizerError> errors)
-        {
-            string errorMessage = default;
-            string errorCode = default;
-
-            if (errors.Count > 0)
-            {
-                var firstError = errors[0];
-
-                errorMessage = firstError.Message;
-                errorCode = firstError.ErrorCode;
-            }
-            else
-            {
-                errorMessage = "Copy model operation failed.";
-            }
-
-            var errorInfo = new Dictionary<string, string>();
-            int index = 0;
-
-            foreach (var error in errors)
-            {
-                errorInfo.Add($"error-{index}", $"{error.ErrorCode}: {error.Message}");
-                index++;
-            }
-
-            return async
-                ? await _diagnostics.CreateRequestFailedExceptionAsync(_response, errorMessage, errorCode, errorInfo).ConfigureAwait(false)
-                : _diagnostics.CreateRequestFailedException(_response, errorMessage, errorCode, errorInfo);
         }
     }
 }